refactor: separate folder merge logic from cli.py to _merge.py
