+++ src/fosslight_source/_help.py
@@ -39,12 +39,12 @@ _HELP_MESSAGE_SOURCE_SCANNER = f"""
     -c <number>            Number of CPU cores/threads to use for scanning
     -t <seconds>           Timeout in seconds for ScanCode scanning
     -j                     Generate raw scanner results in JSON format
+    --no_merge             Keep source paths file-based without folder merge
     --no_correction        Skip OSS information correction with sbom-info.yaml
     --correct_fpath <path> Path to custom sbom-info.yaml file
     --hide_progress        Hide the progress bar during scanning
     --kb_url <url>         KB API URL (priority: parameter > KB_URL env > default)
     --kb_token <token>     KB bearer token (priority: parameter > KB_TOKEN env)
-    --no_merge             Keep source paths file-based without folder merge
 
     💡 Examples
     ────────────────────────────────────────────────────────────────────
